All Towers are now added and fully functional, no need for tests
+ since all piece logic are essantially identical besides the pawn only pawn and one other really needs automatic tests.

+++ src/main/pieces/IPiece.java
@@ -34,7 +34,7 @@ public interface IPiece {
      * Used to get the value of the piece
      * @return
      */
-    public int getValue();
+    public double getValue();
 
     /**
      *
